Modified CI/CD Process Kill script

+++ WebServer/CI/KillServer.sh
@@ -1,3 +1,16 @@
 pid=$(cat /home/pi/PythonPoweredPi/WebServer/CI/PythonPoweredPi.pid)
 echo PreviousPid: $pid
-kill -9 "$pid"
+echo "Attempting to kill previous instance"
+
+if [ "$(ps -o comm= -p "$pid")" = "python3" ]; then 
+	echo "Previous instance found"
+	kill -9 "$pid"
+	echo "Waiting for process to die"
+	sleep 10s
+	echo "Previous instance succesfully shut down."
+else 
+	echo "failed to identify a running instance"
+fi
+
+
+
